+++ index.d.ts
@@ -1,1 +1,3 @@
-export * from './lib';
+export * from './_MouseEvent';
+export * from './trigger_custom_event';
+export * from './trigger_native_event';