+++ trigger_custom_event.d.ts
@@ -1,0 +1,2 @@
+export declare function trigger_custom_event<T>(el: HTMLElement, event_name: string, detail: T): void;
+export { trigger_custom_event as trigger__event__custom };