+++ trigger_custom_event.ts
@@ -1,0 +1,13 @@
+export function trigger_custom_event<T>(el: HTMLElement, event_name: string, detail: T) {
+	let event
+	if (window.CustomEvent) {
+		event = new CustomEvent<T>(event_name, { detail })
+	} else {
+		event = document.createEvent('CustomEvent')
+		event.initCustomEvent(event_name, true, true, detail)
+	}
+	el.dispatchEvent(event)
+}
+export {
+	trigger_custom_event as trigger__event__custom
+}